move IDisposable to tokens
