--- src/main/disposable.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-
-export interface IDisposable {
-  dispose(): Promise<void>;
-}
